journaled doesn't work on Rails 5.1 because of monkey patches
this gem is currently only running tests against rails 5.2 because the
gemspec is set to ~> 5.1. i wrote this test and it was passing on 5.2
but then when i forcibly changed the gemspec to 5.1.x the test no longer
passes.

i believe this is due to how Rails 5.1 implements the eventual delete
invocation after the callbacks are run during a destroy call: it just
calls delete_all https://github.com/rails/rails/blob/5-1-stable/activerecord/lib/active_record/persistence.rb#L527

this implementation changed in 5.2 as you can see here
https://github.com/rails/rails/blob/5-2-stable/activerecord/lib/active_record/persistence.rb#L681
and here
https://github.com/rails/rails/blob/5-2-stable/activerecord/lib/active_record/persistence.rb#L200

h/t Katelyn on those links.

this is a failing test that we oughta fix