Introducing: Tagged Events! (journaled v4.1.0) (#22)
This adds a new `tagged: true` option to `journal_attributes`, and `Journaled.tagged`/`Journaled.tag!` helpers that allow events to be tagged with contextual metadata (useful for audit logs and tracing):

```ruby
class MyEvent
  include Journaled::Event
      
  journal_attributes :attr_1, :attr_2, tagged: true
end

Journaled.tagged(foo: 'bar') do
  # events emitted in this scope will have a `tags` attribute with contents `{ "foo": "bar" }`
end

# events emitted in this scope will have a `tags` attribute with contents `{}`
```

All "tagged" events will be given a `tags` attribute, but it can be empty. Consuming apps with strict schema enforcement (`"additionalProperties": false`) will need to include the "tags" field in any relevant event schemas.

Under the hood, this removes `RequestStore` in favor of `ActiveSupport::CurrentAttributes`. This is the new convention for Rails apps (available in 5.2+), so in theory consuming apps shouldn't be impacted by this, but out of caution we have incremented the gem a minor version (to 4.1.0).