Journaled 5.0: Transactional batching (#25)
This adds transactional batching to journaled. What does that mean? Well, now, by default, when multiple events are emitted inside of a transaction:

```ruby
ActiveRecord::Base.transaction do
  event_1.journal!
  event_2.journal!
end
```

A single job will be enqueued directly before the SQL `COMMIT` statement, batching up the two events. (And if the transaction rolls back, the job will never be enqueued.)

This can be disabled with a global config (for testing purposes -- we should eventually remove this config if we find no downsides in the new behavior):

```ruby
Journaled.transactional_batching_enabled = !Rails.env.production?
```

What happens if we aren't in a transaction? Well, the same thing that happened before! (A job will be enqueued right away.) (**I'm looking into adding an additional transactional safety check, that would raise an error if `journal!` is called outside of a transaction!** But for I'll save that for a future PR.)

Because this bumps the version to 5.0, it also removes compatibility for `Journaled::DeliveryJob` jobs enqueued with legacy (3.1.0-era) arguments. This job now accepts a list of events to emit, rather than a single event-kwarg-blog, so a new legacy input pattern is accepted for compatibility with jobs enqueued by v4.x.x of the gem.