Add `snapshot_on_deletion` feature to AuditLog (#28)
### Summary

Because the `changes` hash is always empty on deletion, we would like the option to create a snapshot by default on deletes, to ensure that we know what it was exactly that was deleted. (This is particularly useful for models that existed prior to the introduction of this framework.)

### Other Information

I'm defaulting it to `false` because ... I guess it felt extra confusing to have a feature that defaults to only partially being enabled. But I could be convinced that `true` is actually a better default 🤷