Add Rails 7.1 testing (#35)
### Summary
I stumbled upon an error when trying to add Rails 7.1 support to a `journaled` dependent: in `audit_log.rb`, we call `skip_audit_log` on certain ActiveRecord classes which no longer inherit from `ActiveRecord::Base` in 7.1, so that would raise with a `NoMethodError`. This PR:

- Bumps `.ruby-version` to 3.2.2 (most if not all of us are using >=3.0) as our daily driver, so it's a bit of an ergonomic change)
- Includes rubocop autocorrect changes (the _vast_ majority of these are the `frozen_string_literal: true` lints; mostly felt the need to do this to unblock running `be rake` locally)
- Skips calling `skip_audit_log` on ActiveRecord classes when on Rails 7.1
- Only sets `legacy_connection_handling` when on Rails 7.0

### Other Information
There might be some other things worth considering (not ignoring lockfiles(?), actually using the new connection handling (even if it is only a dummy app)) but I didn't want to drag out a PR that's already pretty big.